+++ src/goose/_internal/toolkit/prompts/developer.jinja
