+++ src/goose/_internal/toolkit/prompts/github.jinja
