+++ src/goose/_internal/toolkit/prompts/safety_rails.jinja
