+++ src/goose/_internal/toolkit/repo_context/prompts/repo_context.jinja
