+++ src/goose/_internal/toolkit/repo_context/repo_context.py
@@ -7,9 +7,9 @@ from exchange import Message
 
 from goose.notifier import Notifier
 from goose.toolkit import Toolkit
-from goose.toolkit.base import Requirements, tool
-from goose.toolkit.repo_context.utils import get_repo_size, goose_picks_files
-from goose.toolkit.summarization.utils import load_summary_file_if_exists, summarize_files_concurrent
+from goose.toolkit import Requirements, tool
+from .utils import get_repo_size, goose_picks_files
+from ..summarization.utils import load_summary_file_if_exists, summarize_files_concurrent
 from goose.utils.ask import clear_exchange, replace_prompt
