+++ src/goose/_internal/toolkit/summarization/__init__.py
