+++ src/goose/_internal/toolkit/summarization/summarize_project.py
@@ -1,9 +1,8 @@
 import os
 from typing import List, Optional
 
-from goose.toolkit import Toolkit
-from goose.toolkit.base import tool
-from goose.toolkit.summarization.utils import summarize_directory
+from goose.toolkit import Toolkit, tool
+from .utils import summarize_directory
 
 
 class SummarizeProject(Toolkit):
