+++ src/goose/_internal/toolkit/summarization/utils.py
