+++ src/goose/_internal/toolkit/utils.py
