+++ src/goose/cli/main.py
@@ -1,4 +1,3 @@
-import sys
 from datetime import datetime
 from pathlib import Path
 from typing import Dict, Optional
